showtimes overlapping check added

+++ database/migrations/2023_09_17_130915_create_hall_configs_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('hall_configs', function (Blueprint $table) {
             $table->id();
-            $table->integer('hall_id')->unsigned();
+            $table->unsignedBigInteger('hall_id')->unsigned();
             $table->foreign('hall_id')->references('id')->on('halls')->onDelete('cascade');
             $table->integer('row');
             $table->integer('seat');
